Add persistent scene table support to example application
Currently it works like this:

if a database containing the scene table is found, it will be used,
otherwise a new one will be created

each call scene event will try to load the scene value from the database,
if the value is not found the state of the output channel does not
change, except for scene MIN and scene MAX which are hardcoded to
minimum and maximum known values for our device.

INC and DEC scene steps are currently hardcoded as well.

each save scene event will save the current value to the database, using
the scene number as the key

implements #13611